+++ CFQ/Utils/Logger.swift
@@ -3,6 +3,7 @@ import Foundation
 
 enum LogLevel: String {
     case info = "ℹ️ INFO"
+    case action = "💥 ACTION"
     case warning = "⚠️ WARNING"
     case error = "❌ ERROR"
 }
@@ -11,7 +12,7 @@ struct Logger {
     static func log(_ message: String, level: LogLevel = .info, file: String = #file, function: String = #function, line: Int = #line) {
         #if DEBUG
         let filename = (file as NSString).lastPathComponent
-        print("\(level.rawValue) [\(filename):\(line)] \(function) -> \(message)")
+        print("\(level.rawValue) - [\(filename):\(line)] - \(function) -> \(message)")
         #endif
     }
 }
