+++ CFQ/Widget/CardTurn/DatePickerMainInformations.swift
@@ -1,9 +1,3 @@
-//
-//  DatePicker.swift
-//  CFQ
-//
-//  Created by Calvignac Charles on 26/01/2025.
-//
 
 import SwiftUI
 
@@ -33,27 +27,10 @@ struct DatePickerMainInformations: View {
             .sheet(isPresented: $showPicker) {
                 ZStack {
                     NeonBackgroundImage()
-
-                    VStack {
-                        DatePicker("", selection: $viewModel.date, in: Date.now..., displayedComponents: .date)
-                            .datePickerStyle(GraphicalDatePickerStyle())
-                            .environment(\.locale, Locale.init(identifier: "fr_FR"))
-                            .tint(.white)
-                            .colorScheme(.dark)
-                        Button(action: {showPicker = false; hasCompleted = true}, label: {
-                            Text("Done")
-                                .foregroundColor(.white)
-                                .padding(.horizontal, 30)
-                                .padding(.vertical, 10)
-                                .background(.black)
-                                .cornerRadius(8)
-                                .overlay(
-                                    RoundedRectangle(cornerRadius: 8)
-                                        .stroke(Color.white, lineWidth: 0.5)
-                                )
-                            
-                        })
-                    }
+                    CollectionViewDate(viewModel: viewModel, onClose: {
+                        showPicker = false
+                        hasCompleted = true
+                    })
                 }
                 .presentationDragIndicator(.visible)
                 .presentationDetents([.height(500)])
@@ -71,46 +48,8 @@ struct DatePickerMainInformations: View {
                             CollectionViewHours(viewModel: viewModel)
                                 .presentationDragIndicator(.visible)
                                 .presentationDetents([.height(500)])
-                            }
                         }
-                
-                /*
-                DatePicker(
-                            "",
-                            selection: $date,
-                            displayedComponents: .hourAndMinute
-                )
-                .tint(.white)
-                .colorScheme(.dark)
-                .onChange(of: date) { newDate in
-                            let calendar = Calendar.current
-                            let hour = calendar.component(.hour, from: newDate)
-                            var minute = calendar.component(.minute, from: newDate)
-                            let roundedMinute = Int(round(Double(minute) / 15.0)) * 15
-                            
-                            minute = roundedMinute
-                            date = calendar.date(bySettingHour: hour, minute: minute, second: 0, of: newDate) ?? date
-                        }
-                
-                Text("-")
-                    .foregroundColor(.white)
-                DatePicker(
-                            "",
-                            selection: $date,
-                            displayedComponents: .hourAndMinute
-                )
-                .tint(.white)
-                .colorScheme(.dark)
-                .onChange(of: date) { newDate in
-                            let calendar = Calendar.current
-                            let hour = calendar.component(.hour, from: newDate)
-                            var minute = calendar.component(.minute, from: newDate)
-                            let roundedMinute = Int(round(Double(minute) / 15.0)) * 15
-                            
-                            minute = roundedMinute
-                            date = calendar.date(bySettingHour: hour, minute: minute, second: 0, of: newDate) ?? date
-                        }
-                */
+                    }
             }
         }
     }
